+++ src/clevertap.js
@@ -41,6 +41,7 @@ import { hasWebInboxSettingsInLS, checkAndRegisterWebInboxElements, initializeWe
 import { Variable } from './modules/variables/variable'
 import VariableStore from './modules/variables/variableStore'
 import { checkBuilder, addAntiFlicker } from './modules/visualBuilder/pageBuilder'
+import { setServerKey } from './modules/webPushPrompt/prompt'
 
 export default class CleverTap {
   #logger
@@ -509,6 +510,7 @@ export default class CleverTap {
       closeIframe(campaignId, divIdIgnored, this.#session.sessionId)
     }
     api.enableWebPush = (enabled, applicationServerKey) => {
+      setServerKey(applicationServerKey)
       this.notifications._enableWebPush(enabled, applicationServerKey)
     }
     api.tr = (msg) => {
@@ -588,7 +590,7 @@ export default class CleverTap {
       // already initailsed
       return
     }
-    checkBuilder(this.#logger, accountId)
+
     StorageManager.removeCookie('WZRK_P', window.location.hostname)
     if (!this.#account.id) {
       if (!accountId) {
@@ -597,6 +599,7 @@ export default class CleverTap {
       }
       this.#account.id = accountId
     }
+    checkBuilder(this.#logger, this.#account.id)
     this.#session.cookieName = SCOOKIE_PREFIX + '_' + this.#account.id
 
     if (region) {