+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.11.2] - 17 Oct, 2024
+- Fixes a bug where Web Popup Campaign was not working if other iframes present on DOM.
+
+## [1.11.1] - 16 Oct, 2024
+- Fixes a bug where delay was not working for Web Popup Campaigns
+
+## [1.11.0] - 14 Oct, 2024
+- Adds Additional triggers and clicks handling for Web Popup template
+- Adds a fix for account id undefined for visual editor
+- Adds a fix for css in soft prompt
+
+## [1.10.1] - 8 Oct, 2024
+- Adding a fix for the version check in visual editor
+
+## [1.10.0] - 8 Oct, 2024
+- Adds new api to handle rendering of customized web push prompt
+
 ## [1.9.6] - 23 Sept, 2024
 - Shopify support for visual builder